+++ documentation/example/inf/monitor/locals.tf
@@ -3,11 +3,12 @@ locals {
     module.cw_config_ecs.data,
     module.ecs_dashboard.data,
     {
-      alert  = module.alert_account.data
-      api    = module.api_account.data
-      config = module.config_account.data
-      event  = module.event_account.data
-      sms    = module.sms_account.data
+      alert            = module.alert_account.data
+      athena_workgroup = module.athena_workgroup.data
+      api              = module.api_account.data
+      config           = module.config_account.data
+      event            = module.event_account.data
+      sms              = module.sms_account.data
     },
   )
   std_var = {